Move postinstall to src files
